% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cal_pop_wgt}
\alias{cal_pop_wgt}
\title{Calculate population weights using RAF}
\usage{
cal_pop_wgt(
  input_file,
  reference_index_file,
  reference_data_file,
  reference_pop_desc_file
)
}
\arguments{
\item{input_file}{file name of input data containing rsid, chr, bp, a1, a2, and af1}

\item{reference_index_file}{file name of reference panel index data}

\item{reference_data_file}{file name of reference panel data}

\item{reference_pop_desc_file}{file name of reference panel population description data}
}
\value{
R data frame containing population IDs and weights
}
\description{
Calculate population weights using RAF
}
