% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpw2}
\alias{cpw2}
\title{Calculate population weights using RAF}
\usage{
cpw2(
  input_file,
  reference_index_file,
  reference_data_file,
  reference_pop_desc_file,
  interval = NULL
)
}
\arguments{
\item{input_file}{file name of input data containing rsid, chr, bp, a1, a2, and af1}

\item{reference_index_file}{file name of reference panel index data}

\item{reference_data_file}{file name of reference panel data}

\item{reference_pop_desc_file}{file name of reference panel population description data}

\item{interval}{number of non-overlapping SNP sets used in calculating population weights}
}
\value{
R data frame containing population IDs and weights
}
\description{
Calculate population weights using RAF
}
