% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiqt.R
\name{fiqt}
\alias{fiqt}
\title{FDR Inverse Quantile Transformation (FIQT)}
\usage{
fiqt(z, min.p = 10^-300)
}
\arguments{
\item{z}{association Z-score vector}

\item{min.p}{minimum p-value admitted (to avoid zero p-values/adjusted p-values which give troubles with inverse cdf)}
}
\value{
R vector containing the Winner's Curse adjusted Z-scores
}
\description{
FDR Inverse Quantile Transformation (FIQT)
}
