\name{GAUSS-package}
\alias{GAUSS-package}
\alias{GAUSS}
\docType{package}
\title{
\packageTitle{GAUSS}
}
\description{
\packageDescription{GAUSS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GAUSS}
\packageIndices{GAUSS}
Gene-set association analysis using Sparse Signals
}
\author{
\packageAuthor{GAUSS}

Maintainer: \packageMaintainer{GAUSS}
}
\references{
Dutta et al.: A powerful subset-based gene-set analysis method identifies novel associations and improves interpretation in UK Biobank
}
