% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jepeg}
\alias{jepeg}
\title{Joint effect on phenotype of eQTLs/functional SNPs associated with a gene}
\usage{
jepeg(
  study_pop,
  input_file,
  annotation_file,
  reference_index_file,
  reference_data_file,
  reference_pop_desc_file,
  af1_cutoff = NULL
)
}
\arguments{
\item{study_pop}{study population group}

\item{input_file}{file name of GWAS summary statistics data containing rsid, chr, bp, a1, a2 and z}

\item{reference_index_file}{file name of reference panel index data}

\item{reference_data_file}{file name of reference panel data}

\item{reference_pop_desc_file}{file name of reference panel population description data}

\item{af1_cutoff}{cutoff of reference allele, a1, frequency}

\item{annotation}{file name of the SNP annotation data set}
}
\value{
R dataframe containing geneid, chisq, df, jepeg_pval, num_snp, top_categ, top_categ_pval, top_snp, and top_snp_pval
}
\description{
Joint effect on phenotype of eQTLs/functional SNPs associated with a gene
}
