% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGC2_SCZ_ANC_Prop.R
\docType{data}
\name{PGC2_SCZ_ANC_Prop}
\alias{PGC2_SCZ_ANC_Prop}
\title{PGC2 Schizophrenia GWAS Ancestry Proportion}
\format{
An object of class 'data.frame' with 2 columns.
}
\usage{
data(PGC2_SCZ_ANC_Prop)
}
\description{
This data set include two columns with the following names: 
pop (population abbreviation) wgt (weight or proportion). This is estimated 
using the `afmix()` function.
}
\examples{
\dontrun{
data(PGC2_SCZ_ANC_Prop)
str(PGC2_SCZ_ANC_Prop)
}

}
\references{
Schizophrenia Working Group of Psychiatric Genomics Consortium (2014). 
Biological insights from 108 schizophrenia associated genetic loci. Nature
\href{https://www.nature.com/articles/nature13595}{Link to article}
}
\keyword{datasets}
