% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prep_qcat}
\alias{prep_qcat}
\title{Prepare datasets for QCAT analysis}
\usage{
prep_qcat(
  chr,
  start_bp,
  end_bp,
  wing_size,
  study_pop,
  input_file,
  reference_index_file,
  reference_data_file,
  reference_pop_desc_file,
  af1_cutoff = NULL
)
}
\arguments{
\item{chr}{chromosome number}

\item{start_bp}{start base pair position of prediction window}

\item{end_bp}{end base pair position of prediction window}

\item{wing_size}{the size of the area flanking the left and right of the prediction window}

\item{study_pop}{study population group}

\item{input_file}{file name of GWAS summary statistics data containing rsid, chr, bp, a1, a2, af1, and z}

\item{reference_index_file}{file name of reference panel index data}

\item{reference_data_file}{file name of reference panel data}

\item{reference_pop_desc_file}{file name of reference panel population description data}

\item{af1_cutoff}{cutoff of reference allele, a1, frequency}
}
\value{
R dataframe containing rsid, chr, bp, a1, a2, af1ref, z, qcat_chisq, qcat_pval, type
}
\description{
Prepare datasets for QCAT analysis
}
