% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prep_zmix5_sup}
\alias{prep_zmix5_sup}
\title{Calculate population weights using association Z-scores}
\usage{
prep_zmix5_sup(
  input_file,
  reference_index_file,
  reference_data_file,
  reference_pop_desc_file,
  percentile = NULL,
  interval = NULL
)
}
\arguments{
\item{input_file}{file name of input data containing rsid, chr, bp, a1, a2, and z}

\item{reference_index_file}{file name of reference panel index data}

\item{reference_data_file}{file name of reference panel data}

\item{reference_pop_desc_file}{file name of reference panel population description data}

\item{percentile}{percentile cutoff for normalized variance}

\item{interval}{stepping distance within the SNP vector for selecting the first SNP of each pair.}
}
\value{
R data frame containing superpopulation IDs and weights
}
\description{
This function mirrors prep_zmix5 but computes correlations at the
superpopulation level.
}
